user-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1